+++ simLifeDisplayAPI/src/main/java/barsuift/simLife/j3d/tree/TreeBranch3D.java
@@ -21,19 +21,19 @@ package barsuift.simLife.j3d.tree;
 import java.util.List;
 
 import javax.media.j3d.Group;
-import javax.vecmath.Point3d;
-import javax.vecmath.Vector3d;
+import javax.vecmath.Point3f;
+import javax.vecmath.Vector3f;
 
 import barsuift.simLife.Persistent;
 
 public interface TreeBranch3D extends Persistent<TreeBranch3DState> {
 
-    public Point3d getEndPoint();
+    public Point3f getEndPoint();
 
     public List<TreeBranchPart3D> getBranchParts();
 
     public Group getGroup();
 
-    public Vector3d getTranslationVector();
+    public Vector3f getTranslationVector();
 
 }
